rename waiverTypes file
