+++ frontend/src/types/waiverRegistrationTypes.ts
